finance_backend_backup/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
marshmallow==3.21.3
flasgger==0.9.7.1
python-dotenv==1.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.30
